+++ iceoryx_posh/include/iceoryx_posh/internal/popo/building_blocks/chunk_receiver.hpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2020 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
+// Copyright (c) 2021 - 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -33,6 +33,23 @@ enum class ChunkReceiveResult
     TOO_MANY_CHUNKS_HELD_IN_PARALLEL,
     NO_CHUNK_AVAILABLE
 };
+
+/// @brief Converts the ChunkReceiveResult to a string literal
+/// @param[in] value to convert to a string literal
+/// @return pointer to a string literal
+inline constexpr const char* asStringLiteral(const ChunkReceiveResult value) noexcept;
+
+/// @brief Convenience stream operator to easily use the `asStringLiteral` function with std::ostream
+/// @param[in] stream sink to write the message to
+/// @param[in] value to convert to a string literal
+/// @return the reference to `stream` which was provided as input parameter
+inline std::ostream& operator<<(std::ostream& stream, ChunkReceiveResult value) noexcept;
+
+/// @brief Convenience stream operator to easily use the `asStringLiteral` function with iox::log::LogStream
+/// @param[in] stream sink to write the message to
+/// @param[in] value to convert to a string literal
+/// @return the reference to `stream` which was provided as input parameter
+inline log::LogStream& operator<<(log::LogStream& stream, ChunkReceiveResult value) noexcept;
 
 /// @brief The ChunkReceiver is a building block of the shared memory communication infrastructure. It extends
 /// the functionality of a ChunkQueuePopper with the abililty to pass chunks to the user side (user process).